refactor : role 도메인 > member 도메인과 통합, enum 클래스 패키지 통합
