+++ src/main/java/org/sugar_square/community_service/domain/report/Report.java
@@ -16,6 +16,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.sugar_square.community_service.domain.BaseEntity;
 import org.sugar_square.community_service.domain.member.Member;
+import org.sugar_square.community_service.enums.ErrorCode;
+import org.sugar_square.community_service.enums.ErrorGroup;
 
 @Entity
 @Getter
