--- src/main/java/org/sugar_square/community_service/domain/role/UserRolePK.java
@@ -1,29 +1,0 @@
-package org.sugar_square.community_service.domain.role;
-
-import jakarta.persistence.Embeddable;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import java.io.Serializable;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.sugar_square.community_service.domain.member.Member;
-
-@Embeddable
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-@EqualsAndHashCode
-public class UserRolePK implements Serializable {
-
-  @ManyToOne(fetch = FetchType.LAZY)
-  @JoinColumn(name = "member_id", nullable = false, updatable = false)
-  private Member member; // TODO: DTO 에서 한 번 더 null 검증
-
-  @ManyToOne(fetch = FetchType.LAZY)
-  @JoinColumn(name = "role_id", nullable = false, updatable = false)
-  private Role role; // TODO: DTO 에서 한 번 더 null 검증
-}
