+++ src/test/java/org/sugar_square/community_service/BaseRepositoryTest.java
@@ -1,6 +1,7 @@
 package org.sugar_square.community_service;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.sugar_square.community_service.enums.RoleEnum.USER;
 
 import java.util.Optional;
 import org.junit.jupiter.api.BeforeEach;
@@ -30,6 +31,7 @@ public class BaseRepositoryTest {
             .username("test")
             .password("test")
             .nickname("test")
+            .role(USER)
             .build()
     );
   }
